Compile and output CSS file for CSS link to work
It was just due to CSS loader. Although it produces no warnings to place
MiniCSSExtractPlugin at the beginning of the loader order, it results in
the CSS not being output. Putting it in the middle results in warnings.
However deleting the style-loader and simply having the css loader and
the plugin as shown in various guides, works with no errors. SO maybe
the style-loader is redundant.